perf: improve query data iteration (#6)
* fix: pointer iteration for bigger values

* fix: remove fieldhash variable

* fix: reduced size of clause state

* fix: levenshtein for keyword conditions always apply

* fix: localized token variable

* fix: simplified tokens reset on iterations

* fix: reduced the number of calls to handle per iteration

* fix: reduced the number of handle calls in keyword iteration

* fix: only include scores greater than zero

* fix: faster resolution of candidates

* fix: only resolve when the bitmap is not empty

* fix: do not check bounds

* fix: moved to float32 for faster calculation

* doc: updated documentation

* fix: pending is ensuring the arriving entries are always not set

* fix: do io heavy operations near each other

* fix: score compared against contant instead

* fix: improvement of performance by early checking the state is empty

* fix: use of agner fog bucket grouping of math operations

* fix: allow flexible delta

* fix: included original delta for comparison

* fix: fixed fastlog.Ln to safely handle bounds of float32

* chore(release): 1.7.4

+++ dorks/dorks_test.go
@@ -187,7 +187,7 @@ func TestParse(t *testing.T) {
 					{
 						Keyword: "age",
 						Match:   &dorks.Match{Operator: dorks.MatchOperatorGreaterEqual, Value: "30"},
-						Boost:   new(30.0),
+						Boost:   new(float32(30.0)),
 					},
 				}}},
 			{Query: "+Antonio Donis age:<=30 +country:Colombia", Expect: dorks.Query{Dorks: []*dorks.Dork{
